Refactor styles to separate concerns better. Apply overrides more cleanly.
